feat: dynamic route 적용, state 재 설정
- dynamic route로 각 픽업 페이지 이동
- 각 픽업 별 이상 state 유지
