+++ components/gacha/GachaControls.vue
@@ -2,7 +2,6 @@
   <div class="controls">
     <button @click="$emit('drawSingle')">1회 추출</button>
     <button @click="$emit('drawTen')">10회 추출</button>
-    <button @click="resetDrawCount()">초기화</button>
   </div>
 </template>
 
@@ -12,9 +11,6 @@ import { mapMutations } from "vuex";
 
 export default Vue.extend({
   name: "GachaControls",
-  methods: {
-    ...mapMutations(["resetDrawCount"]),
-  },
 });
 </script>
